+++ changelog.md
@@ -1,6 +1,27 @@
+# 3.8.0
+
+## Changes
+
+Updated for patch 10.1.0.
+
+Added token and trinket data from Aberrus the Shadowed Crucible
+
+### Require note per response
+
+It's now possible to require notes per specific response - check `Buttons and Response` options and check `Require Notes` for each response you want raiders to submit a note with.  
+Old require notes option has been removed as part of this change.
+
+Note: This will not work for people with older versions.
+
+### Void-Touched Curio
+
+Removed from blacklist.
+
 ## Bugfixes
 
 - *Fixed occasional inability to trade multiple identical items.*
+- *`/rc fulltest` should once again pull items from the latest raid.*
+
 
 # 3.7.1